requirements.txt
# AgentBridge Python dependencies

# HTTP client (Phase 1)
requests>=2.28.0

# gRPC client (Phase 2+)
grpcio>=1.50.0
protobuf>=4.21.0

# Note: The generated protobuf stubs (AgentBridgeServer, TempoScripting)
# are installed by Tempo's build process into the TempoEnv virtual environment.
# The MCP server auto-detects the Tempo API path, or you can set TEMPO_API_PATH:
#   export TEMPO_API_PATH=/path/to/YourProject/Plugins/Tempo/TempoCore/Content/Python/API/tempo
